drivers/net: fix port id size
The ethdev port id should be 16 bits now. This patch changes the
variable size of port id in some net drivers from 8 bits to 16
bits.

Fixes: 09b23f8b9df6 ("net/bnxt: fix port stop process and cleanup resources")
Fixes: 769de16872ab ("net/bnxt: fix port default rule create/destroy")
Fixes: 50370662b727 ("net/ice: support device and queue ops")
Fixes: a50d7cbbdad7 ("net/qede: support registers dump")
Cc: stable@dpdk.org

Acked-by: Ajit Khaparde <ajit.khaparde@broadcom.com>

+++ drivers/net/bnxt/tf_ulp/bnxt_ulp.c
@@ -646,7 +646,7 @@ static void
 bnxt_ulp_destroy_vfr_default_rules(struct bnxt *bp, bool global)
 {
 	struct bnxt_ulp_vfr_rule_info *info;
-	uint8_t port_id;
+	uint16_t port_id;
 	struct rte_eth_dev *vfr_eth_dev;
 	struct bnxt_representor *vfr_bp;
