+++ drivers/net/qede/base/bcm_osal.h
@@ -382,7 +382,7 @@ void qede_hw_err_notify(struct ecore_hwfn *p_hwfn,
 /* TODO: */
 #define OSAL_SCHEDULE_RECOVERY_HANDLER(hwfn) nothing
 
-int qede_save_fw_dump(uint8_t port_id);
+int qede_save_fw_dump(uint16_t port_id);
 
 #define OSAL_SAVE_FW_DUMP(port_id) qede_save_fw_dump(port_id)
